+++ src/solvers/dgsem/calc_volume_integral.jl
@@ -190,26 +190,6 @@ function calc_volume_integral!(backend::Nothing, du, u, mesh,
     return nothing
 end
 
-function calc_volume_integral!(backend::Backend, du, u, mesh,
-                               have_nonconservative_terms, equations,
-                               volume_integral, dg::DGSEM, cache)
-    nelements(dg, cache) == 0 && return nothing
-    kernel! = volume_integral_KAkernel!(backend)
-    kernel_cache = kernel_filter_cache(cache)
-    kernel!(du, u, typeof(mesh), have_nonconservative_terms, equations,
-            volume_integral, dg, kernel_cache,
-            ndrange = nelements(dg, cache))
-    return nothing
-end
-
-@kernel function volume_integral_KAkernel!(du, u, MeshT,
-                                           have_nonconservative_terms, equations,
-                                           volume_integral, dg::DGSEM, cache)
-    element = @index(Global)
-    volume_integral_kernel!(du, u, element, MeshT, have_nonconservative_terms,
-                            equations, volume_integral, dg, cache)
-end
-
 @inline function calc_volume_integral!(backend::Nothing, du, u, mesh,
                                        have_nonconservative_terms, equations,
                                        volume_integral::VolumeIntegralAdaptive{<:IndicatorHennemannGassner},