+++ test/test_tree_1d_euler.jl
@@ -393,6 +393,15 @@ end
     @test_allocations(Trixi.rhs!, semi, sol, 1000)
 end
 
+@trixi_testset "elixir_euler_leblanc_limiter_liu_zhang.jl" begin
+    @test_trixi_include(joinpath(EXAMPLES_DIR,
+                                 "elixir_euler_leblanc_limiter_liu_zhang.jl"),
+                        l2=[0.212900601, 0.0596467559, 0.0249364541],
+                        linf=[0.577864839, 0.105468742, 0.0630698303])
+    @test length(global_limiter!.history_davis_yin_iterations) == 1
+    @test_allocations(Trixi.rhs!, semi, sol, 1000)
+end
+
 @trixi_testset "elixir_euler_blast_wave.jl" begin
     @test_trixi_include(joinpath(EXAMPLES_DIR, "elixir_euler_blast_wave.jl"),
                         l2=[0.21934822867340323, 0.28131919126002686, 0.554361702716662],